brand-media-platform-be/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
channels==4.3.2
channels_redis==4.3.0
Django==5.2.10
django-ckeditor==6.7.3
django-filter==25.2
django-js-asset==3.1.2
django-parler==2.3
django-parler-rest==2.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
redis==7.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
